Debugging ECDSA 2
